requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
sqlalchemy
psycopg2-binary
alembic